requirements.txt
asgiref==3.5.2
certifi==2022.5.18.1
distlib==0.3.4
Django==4.0.5
django-cors-headers==3.13.0
djangorestframework==3.13.1
filelock==3.7.1
Pillow==9.1.1
pipenv==2022.6.7
platformdirs==2.5.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
virtualenv==20.14.1
virtualenv-clone==0.5.7
